requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0
Pillow==10.0.1
Flask-Cors==4.0.0
Werkzeug==2.3.7
requests==2.31.0
psutil==5.9.5
gunicorn==23.0.0
